Add missing gs-gwc.xml initialization for pgconfig
